+++ services/diagnostic/handlers.go
@@ -17,6 +17,7 @@ import (
 	"github.com/influxdata/kapacitor/models"
 	alertservice "github.com/influxdata/kapacitor/services/alert"
 	"github.com/influxdata/kapacitor/services/alerta"
+	"github.com/influxdata/kapacitor/services/discord"
 	"github.com/influxdata/kapacitor/services/ec2"
 	"github.com/influxdata/kapacitor/services/hipchat"
 	"github.com/influxdata/kapacitor/services/httppost"
@@ -594,6 +595,32 @@ func (h *SlackHandler) WithContext(ctx ...keyvalue.T) slack.Diagnostic {
 	fields := logFieldsFromContext(ctx)
 
 	return &SlackHandler{
+		l: h.l.With(fields...),
+	}
+}
+
+// Discord Handler
+
+type DiscordHandler struct {
+	l Logger
+}
+
+func (h *DiscordHandler) InsecureSkipVerify() {
+	h.l.Info("service is configured to skip ssl verification")
+}
+
+func (h *DiscordHandler) Error(msg string, err error) {
+	h.l.Error(msg, Error(err))
+}
+
+func (h *DiscordHandler) TemplateError(err error, kv keyvalue.T) {
+	h.l.Error("failed to evaluate Discord template", Error(err), String(kv.Key, kv.Value))
+}
+
+func (h *DiscordHandler) WithContext(ctx ...keyvalue.T) discord.Diagnostic {
+	fields := logFieldsFromContext(ctx)
+
+	return &DiscordHandler{
 		l: h.l.With(fields...),
 	}
 }